+++ packages/tide_kit/lib/src/services/tide_keybinding_service.dart
@@ -34,11 +34,6 @@ class TideKeybindingService {
     }
     return shortcuts;
   }
-
-  /// Called whenever this widget receives a keyboard event.
-  void onKeyEvent(KeyEvent event) {
-    Tide.log('Tide: TideKeybindingService key event: ${event.logicalKey}');
-  }
 }
 
 class TideActivateIntent extends ActivateIntent {