+++ packages/tide_kit/lib/src/status_bar/tide_status_bar.dart
@@ -52,6 +52,7 @@ class TideStatusBar extends StatelessWidget {
             // TODO: make these not hardcoded here,
             if (item is TideStatusBarItemText) {
               return TideStatusBarItemTextWidget(
+                icon: item.icon,
                 item: item,
                 position: item.position,
                 text: item.text,