+++ packages/tide_kit/lib/src/tide_core.dart
@@ -1,4 +1,6 @@
 import 'package:equatable/equatable.dart';
+import 'package:flutter/services.dart';
+import 'package:flutter/widgets.dart';
 import 'package:get_it/get_it.dart';
 import 'package:uuid/uuid.dart';
 
@@ -34,3 +36,9 @@ class TideId extends Equatable {
 class TideRegistry {}
 
 const _uuid = Uuid();
+
+class TideLogicalKeyIntent extends Intent {
+  const TideLogicalKeyIntent(this.key);
+
+  final LogicalKeyboardKey key;
+}