+++ packages/tide_kit/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.3.0
+
+- Added [TideQuickInputBox] as a way to gather user input from a [TextField].
+- Added [TideQuickPick] as a way to gather user input from a list of items.
+
+- The onPressed for TideStatusBarItem now contains the BuildContext.
+
 ## 1.2.0
 
 - Activity bar items: Added selectable option which defaults to true, and highlight on hover.